Improvement: customers test
